form-filler/pyproject.toml
[project]
name = "form-filler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "baml-py>=0.67.0",
    "prettytable>=3.12.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "termcolor>=2.5.0",
]
